+++ src/smploaders/ft2_load_aiff.c
@@ -588,7 +588,7 @@ static double getAIFFSampleRate(uint8_t *in)
 	const uint64_t mantissa63 = SWAP64(mantissaBits) & INT64_MAX;
 
 	double dExp = exp15 - EXP_BIAS;
-	double dMantissa = mantissa63 / (INT64_MAX+1.0);
+	double dMantissa = mantissa63 / ((double)INT64_MAX+1.0);
 
 	return (1.0 + dMantissa) * exp2(dExp);
 }
